Updated exporter to include category mappings, and updated models to use Decimal instead of float

+++ src/entropia_skillscanner/models.py
@@ -1,6 +1,8 @@
+# models.py
 from __future__ import annotations
 
 from dataclasses import dataclass
+from decimal import Decimal
 
 
 @dataclass(frozen=True)
@@ -8,7 +10,7 @@ class SkillRow:
     """Normalized representation of a single extracted skill entry."""
 
     name: str
-    value: float
+    value: float          # pipeline/UI-derived (float is fine here)
     added: str
 
 
@@ -17,12 +19,11 @@ class ExportSkill:
     """Skill ready for CSV emission (with category attached)."""
 
     name: str
-    value: float
+    value: Decimal        # export uses Decimal
     category: str
 
 
 @dataclass(frozen=True)
 class ExportTotals:
     category: str
-    total: float
-
+    total: Decimal        # export uses Decimal
